+++ src/index.js
@@ -1,2 +1,3 @@
 import MemoryAllocator from './MemoryAllocator.js';
 export default MemoryAllocator;
+export { default as MemoryRef } from './MemoryRef.js';