+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Export `MemoryRef` (...).
+
 ## [0.0.3] - 2023-10-07
 
 ### Fixed